Added File Creation/Deletion Prank duckyscript
This script includes a little bit of privilege escalation for the PowerShell so that it can modify certain files. This inject.bin in particular is harmless as it simply creates a TacoTuesday folder in the /Users directory. Alternatively, this can be used to delete important files or make modifications to higher-level folders.